web: Common: Refactoring of AdvancedSelector v2
